% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotavolume.R
\name{cotavolume}
\alias{cotavolume}
\title{cotavolume}
\usage{
cotavolume(ini, fim, passo = 0.01, c0 = 0, c1 = 0, c2 = 0, c3 = 0, c4 = 0)
}
\arguments{
\item{ini}{valor da cota inicial da curva (obrigatório)}

\item{fim}{valor da cota final da curva (obrigatório)}

\item{passo}{passo de cota da curva (obrigatório)}

\item{c0}{valor da coeficiente do gator de expoente 0 da curva (opcional)}

\item{c1}{valor da coeficiente do gator de expoente 1 da curva (opcional)}

\item{c2}{valor da coeficiente do gator de expoente 2 da curva (opcional)}

\item{c3}{valor da coeficiente do gator de expoente 3 da curva (opcional)}

\item{c4}{valor da coeficiente do gator de expoente 4 da curva (opcional)}
}
\value{
a função retorna um data frame com as duas colunas, cota e volume,
calculcados com o passo e intervalo de cota definidos pelo usuário na função.
}
\description{
Gera a curva cota-volume de um reservatório por meio de um polinômio do tipo
co + c1 x + c2 x^2 + c3 x^3 + c4  x^4
}
